+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fix
 - errored or skipped referenced httpregion also skips current http region (AnWeber/httpyac#842, #339)
 - failed requests did not update test runner indicator (#348)
+- override of script console in test runner does not prevent logging in output channel (#347)
 
 ## [6.16.4] (2024-11-03)